Add files via upload
This file contains the code used in the arduino for the programming of different electrical equipments used in the manufacturing of our steam materials- Water Cycle demonstration. The code run in such a way that it provides the signal to the relay for the oscillation of the L.E.D motors, and also to run the water motor in periodic interval of 4 seconds.